[refactor] Move code away from lang_utils (#2209)
* [refactor] Move code away from lang_utils

* [skip ci] enforce code format

* fix
