+++ taichi/ir/expression.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "taichi/ir/ir.h"
 #include "taichi/ir/expr.h"
 
 TLANG_NAMESPACE_BEGIN
