+++ taichi/ir/expression_ops.cpp
@@ -1,0 +1,9 @@
+#include "taichi/ir/frontend.h"
+
+namespace taichi {
+namespace lang {
+#define TI_EXPRESSION_IMPLEMENTATION
+#include "taichi/ir/expression_ops.h"
+
+}  // namespace lang
+}  // namespace taichi
