+++ taichi/ir/ir_modified.h
@@ -1,0 +1,9 @@
+#pragma once
+
+namespace taichi {
+namespace lang {
+
+class IRModified {};
+
+}  // namespace lang
+}  // namespace taichi
