+++ taichi/ir/state_machine.cpp
@@ -2,6 +2,7 @@
 
 #include "taichi/ir/statements.h"
 #include "taichi/ir/analysis.h"
+#include "taichi/ir/ir_modified.h"
 
 TLANG_NAMESPACE_BEGIN
