+++ taichi/ir/type.h
@@ -320,4 +320,99 @@ class BitArrayType : public Type {
   int element_num_bits_;
 };
 
+class TypedConstant {
+ public:
+  DataType dt;
+  union {
+    uint64 value_bits;
+    int32 val_i32;
+    float32 val_f32;
+    int64 val_i64;
+    float64 val_f64;
+    int8 val_i8;
+    int16 val_i16;
+    uint8 val_u8;
+    uint16 val_u16;
+    uint32 val_u32;
+    uint64 val_u64;
+  };
+
+ public:
+  TypedConstant() : dt(PrimitiveType::unknown) {
+  }
+
+  TypedConstant(DataType dt) : dt(dt) {
+    value_bits = 0;
+  }
+
+  TypedConstant(int32 x) : dt(PrimitiveType::i32), val_i32(x) {
+  }
+
+  TypedConstant(float32 x) : dt(PrimitiveType::f32), val_f32(x) {
+  }
+
+  TypedConstant(int64 x) : dt(PrimitiveType::i64), val_i64(x) {
+  }
+
+  TypedConstant(float64 x) : dt(PrimitiveType::f64), val_f64(x) {
+  }
+
+  template <typename T>
+  TypedConstant(DataType dt, const T &value) : dt(dt) {
+    // TODO: loud failure on pointers
+    dt.set_is_pointer(false);
+    if (dt->is_primitive(PrimitiveTypeID::f32)) {
+      val_f32 = value;
+    } else if (dt->is_primitive(PrimitiveTypeID::i32)) {
+      val_i32 = value;
+    } else if (dt->is_primitive(PrimitiveTypeID::i64)) {
+      val_i64 = value;
+    } else if (dt->is_primitive(PrimitiveTypeID::f64)) {
+      val_f64 = value;
+    } else if (dt->is_primitive(PrimitiveTypeID::i8)) {
+      val_i8 = value;
+    } else if (dt->is_primitive(PrimitiveTypeID::i16)) {
+      val_i16 = value;
+    } else if (dt->is_primitive(PrimitiveTypeID::u8)) {
+      val_u8 = value;
+    } else if (dt->is_primitive(PrimitiveTypeID::u16)) {
+      val_u16 = value;
+    } else if (dt->is_primitive(PrimitiveTypeID::u32)) {
+      val_u32 = value;
+    } else if (dt->is_primitive(PrimitiveTypeID::u64)) {
+      val_u64 = value;
+    } else {
+      TI_NOT_IMPLEMENTED
+    }
+  }
+
+  template <typename T>
+  bool equal_value(const T &value) const {
+    return equal_type_and_value(TypedConstant(dt, value));
+  }
+
+  std::string stringify() const;
+
+  bool equal_type_and_value(const TypedConstant &o) const;
+
+  bool operator==(const TypedConstant &o) const {
+    return equal_type_and_value(o);
+  }
+
+  int32 &val_int32();
+  float32 &val_float32();
+  int64 &val_int64();
+  float64 &val_float64();
+  int8 &val_int8();
+  int16 &val_int16();
+  uint8 &val_uint8();
+  uint16 &val_uint16();
+  uint32 &val_uint32();
+  uint64 &val_uint64();
+  int64 val_int() const;
+  uint64 val_uint() const;
+  float64 val_float() const;
+  float64 val_cast_to_float64() const;
+};
+
 TLANG_NAMESPACE_END
