+++ taichi/ir/type_utils.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include "taichi/ir/type.h"
 
 namespace taichi {
