+++ taichi/lang_util.h
@@ -4,6 +4,8 @@
 #include "taichi/util/io.h"
 #include "taichi/common/core.h"
 #include "taichi/system/profiler.h"
+#include "taichi/ir/ir_modified.h"
+#include "taichi/ir/stmt_op_types.h"
 #include "taichi/ir/type.h"
 #include "taichi/ir/type_utils.h"
 #include "taichi/ir/type_factory.h"
@@ -21,180 +23,6 @@ real measure_cpe(std::function<void()> target,
 struct Context;
 
 using FunctionType = std::function<void(Context &)>;
-
-enum class SNodeType {
-#define PER_SNODE(x) x,
-#include "taichi/inc/snodes.inc.h"
-#undef PER_SNODE
-};
-
-std::string snode_type_name(SNodeType t);
-
-bool is_gc_able(SNodeType t);
-
-enum class UnaryOpType : int {
-#define PER_UNARY_OP(x) x,
-#include "taichi/inc/unary_op.inc.h"
-#undef PER_UNARY_OP
-};
-
-std::string unary_op_type_name(UnaryOpType type);
-
-inline bool constexpr unary_op_is_cast(UnaryOpType op) {
-  return op == UnaryOpType::cast_value || op == UnaryOpType::cast_bits;
-}
-
-inline bool constexpr is_trigonometric(UnaryOpType op) {
-  return op == UnaryOpType::sin || op == UnaryOpType::asin ||
-         op == UnaryOpType::cos || op == UnaryOpType::acos ||
-         op == UnaryOpType::tan || op == UnaryOpType::tanh;
-}
-
-// Regular binary ops:
-// Operations that take two operands, and returns a single operand with the
-// same type
-
-enum class BinaryOpType : int {
-#define PER_BINARY_OP(x) x,
-#include "inc/binary_op.inc.h"
-#undef PER_BINARY_OP
-};
-
-inline bool binary_is_bitwise(BinaryOpType t) {
-  return t == BinaryOpType ::bit_and || t == BinaryOpType ::bit_or ||
-         t == BinaryOpType ::bit_xor || t == BinaryOpType ::bit_shl ||
-         t == BinaryOpType ::bit_sar;
-}
-
-std::string binary_op_type_name(BinaryOpType type);
-
-inline bool is_comparison(BinaryOpType type) {
-  return starts_with(binary_op_type_name(type), "cmp");
-}
-
-inline bool is_bit_op(BinaryOpType type) {
-  return starts_with(binary_op_type_name(type), "bit");
-}
-
-std::string binary_op_type_symbol(BinaryOpType type);
-
-enum class TernaryOpType : int { select, undefined };
-
-std::string ternary_type_name(TernaryOpType type);
-
-enum class AtomicOpType : int { add, sub, max, min, bit_and, bit_or, bit_xor };
-
-std::string atomic_op_type_name(AtomicOpType type);
-BinaryOpType atomic_to_binary_op_type(AtomicOpType type);
-
-enum class SNodeOpType : int {
-  is_active,
-  length,
-  activate,
-  deactivate,
-  append,
-  clear,
-  undefined
-};
-
-std::string snode_op_type_name(SNodeOpType type);
-
-class IRModified {};
-
-class TypedConstant {
- public:
-  DataType dt;
-  union {
-    uint64 value_bits;
-    int32 val_i32;
-    float32 val_f32;
-    int64 val_i64;
-    float64 val_f64;
-    int8 val_i8;
-    int16 val_i16;
-    uint8 val_u8;
-    uint16 val_u16;
-    uint32 val_u32;
-    uint64 val_u64;
-  };
-
- public:
-  TypedConstant() : dt(PrimitiveType::unknown) {
-  }
-
-  TypedConstant(DataType dt) : dt(dt) {
-    value_bits = 0;
-  }
-
-  TypedConstant(int32 x) : dt(PrimitiveType::i32), val_i32(x) {
-  }
-
-  TypedConstant(float32 x) : dt(PrimitiveType::f32), val_f32(x) {
-  }
-
-  TypedConstant(int64 x) : dt(PrimitiveType::i64), val_i64(x) {
-  }
-
-  TypedConstant(float64 x) : dt(PrimitiveType::f64), val_f64(x) {
-  }
-
-  template <typename T>
-  TypedConstant(DataType dt, const T &value) : dt(dt) {
-    // TODO: loud failure on pointers
-    dt.set_is_pointer(false);
-    if (dt->is_primitive(PrimitiveTypeID::f32)) {
-      val_f32 = value;
-    } else if (dt->is_primitive(PrimitiveTypeID::i32)) {
-      val_i32 = value;
-    } else if (dt->is_primitive(PrimitiveTypeID::i64)) {
-      val_i64 = value;
-    } else if (dt->is_primitive(PrimitiveTypeID::f64)) {
-      val_f64 = value;
-    } else if (dt->is_primitive(PrimitiveTypeID::i8)) {
-      val_i8 = value;
-    } else if (dt->is_primitive(PrimitiveTypeID::i16)) {
-      val_i16 = value;
-    } else if (dt->is_primitive(PrimitiveTypeID::u8)) {
-      val_u8 = value;
-    } else if (dt->is_primitive(PrimitiveTypeID::u16)) {
-      val_u16 = value;
-    } else if (dt->is_primitive(PrimitiveTypeID::u32)) {
-      val_u32 = value;
-    } else if (dt->is_primitive(PrimitiveTypeID::u64)) {
-      val_u64 = value;
-    } else {
-      TI_NOT_IMPLEMENTED
-    }
-  }
-
-  template <typename T>
-  bool equal_value(const T &value) const {
-    return equal_type_and_value(TypedConstant(dt, value));
-  }
-
-  std::string stringify() const;
-
-  bool equal_type_and_value(const TypedConstant &o) const;
-
-  bool operator==(const TypedConstant &o) const {
-    return equal_type_and_value(o);
-  }
-
-  int32 &val_int32();
-  float32 &val_float32();
-  int64 &val_int64();
-  float64 &val_float64();
-  int8 &val_int8();
-  int16 &val_int16();
-  uint8 &val_uint8();
-  uint16 &val_uint16();
-  uint32 &val_uint32();
-  uint64 &val_uint64();
-  int64 val_int() const;
-  uint64 val_uint() const;
-  float64 val_float() const;
-  float64 val_cast_to_float64() const;
-};
 
 inline std::string make_list(const std::vector<std::string> &data,
                              std::string bracket = "") {
