+++ taichi/program/arch.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <string>
-#include "taichi/lang_util.h"
+#include "taichi/common/core.h"
 
 TLANG_NAMESPACE_BEGIN
