+++ taichi/struct/struct.cpp
@@ -2,6 +2,7 @@
 
 #include "taichi/ir/ir.h"
 #include "taichi/ir/expression.h"
+#include "taichi/program/snode_expr_utils.h"
 #include "taichi/program/program.h"
 #include "struct.h"
