+++ taichi/transforms/die.cpp
@@ -4,6 +4,8 @@
 #include "taichi/ir/statements.h"
 #include "taichi/ir/transforms.h"
 #include "taichi/ir/visitors.h"
+#include "taichi/system/profiler.h"
+
 #include <unordered_set>
 
 TLANG_NAMESPACE_BEGIN
