+++ taichi/transforms/optimize_bit_struct_stores.cpp
@@ -3,6 +3,7 @@
 #include "taichi/ir/statements.h"
 #include "taichi/ir/transforms.h"
 #include "taichi/ir/visitors.h"
+#include "taichi/program/compile_config.h"
 
 namespace {
