+++ taichi/transforms/remove_loop_unique.cpp
@@ -2,6 +2,7 @@
 #include "taichi/ir/statements.h"
 #include "taichi/ir/transforms.h"
 #include "taichi/ir/visitors.h"
+#include "taichi/system/profiler.h"
 
 TLANG_NAMESPACE_BEGIN
