+++ taichi/transforms/reverse_segments.cpp
@@ -2,6 +2,7 @@
 #include "taichi/ir/transforms.h"
 #include "taichi/ir/visitors.h"
 #include "taichi/ir/frontend_ir.h"
+#include "taichi/system/profiler.h"
 
 #include <set>
